helm: Use gk chart repo instead of local file
Use the published charts to avoid "it works for me" kind of things
during CI.

Testing chart changes is going to be a bit more difficult though. In
most cases I believe the charts should need to be backwards compatible
anyway, so new changes sould be ok. May wind up making an "unstable"
repo as well that comes and goes to avoid containmenating stable