refactor: Remove an internal serialization format
Remove the internal serialization format of name -> spec. This was used
in 1 location, and didn't benefit from being a map.

In earlier versions this was the returned-to-the-api datastructure, when
there was only booleans. Now that there is more, the name -> spec map
isn't valuable. The whole list is processed anyway.

If there were to be a reason for finding particular features quickly, by
name, a better solution would be to get it from ets directly and process
only the needed values.