collapse user values
Basic test/case for collapsing user values.

Includes a "default" user spec of undefined for the time being to be
used partially to trigger function pattern matching for when the user
isn't defined.

The user can't be defined yet in a deployed version of this code, but
we'll pretend it's there for the time being.