ui: Don't show complex specs if bool true
Don't show the non-boolean spec types if the boolean is true. The data
will still be store, so unchecking will bring back the data. This is to
show in the UI more how the logic works with evaluation.

If boolean was true, but other specs were visible then it would be
difficult to figure out what was meant to happen.

Now with the boolean true disabling the other things, this becomes more
obvious that the other specs are irrelevant.