storage: Move to behavior wrapper
Instead of having the calling module also have the callback definitions,
add a lib module that is the wrapper as well as the behaviour.

This is similar to how gen_server is the behaviour and the module to
call to interact with gen_servers.

Still needs tests around the lib bit