analytics: Add ensure_goal for internal API
Add options for adding a goal that pass ensure_goal to make sure that
the event is a goal before adding the user.

This currently forces a slow path through the gen_server router process
to add the goal. It could check the registration first. If it's
registered as a goal, then adding the goal won't be needed.