analytics: Daemonset 404s on GET
The daemonset previously called the router for GET, but the router isn't
setup in daemonset mode. This should 404 instead of explode.

Also clean up 404s that aren't used. Either the Swagger should
invalidate the data, or we have something unhandled in the code. In
which case this should be an error in our code.