api: Handle un-specified methods
This would previously crash and 500 instead of returning a more helpful
error message

Also wrapped into this :/ is a change in how the test setup is run. This
moves the load/unload functions into a test set so that unload is
actually run. During the dev of this change the UX of running tests was
bad as all the remaining tests were failing