features_store: Write to store after starting
To handle the case where there is no daemonset configmap written yet.
This is a bit messy in that the write for the DS is tied to the general
storing of things, but this is typically not a billable API call so
shouldn't add any cost to do this write.