counter: Don't save for a second identical write
This was previously decided to persist on any write. Now only persist if
the counter sees this as a new write. This will lower API costs (fewer
writes) and improve performance (fewer writes).

The benefits for this are likely only needed for the use case where
writes to the same counter occur frequently, but only to the same
counters. Counters that were un-used or frequently had new users won't
wind up benefiting.