router: Refactor getting counters
Move out the logic of what counters to use from using them.

This will start to allow changing the number, and which, counters to use
when an event is added.

This is a little messy in that the name of the counter and the
registration are returned, it would be nice to return only the
registration with/without a pid, or ensure they are started before
returning which counters to use.