bloom_filter: Rename
Rename to make this more of a counter config.

The counter config used for bloom filters is (UX-wise) a great place to
put in date-cohort information/enabling. This would make the module a
weird mix of config used in various places, but I plan to deal with that
at a later time when the module takes a more defined shape with the
additional counter config