Weekly Cohort Counters
Add counters that are automatically generated by week number. This
allows tracking number of events over time and can be useful for
tracking things that happen multiple times with the same user.

Given that incoming events don't have a particular date to them, we
assume they are "now", there is not a way to backfill with proper week
events