router: Only checks keys for named counters
Filter counters by id type before checking the counters for a key. This
removes the global_counter in the current case, as well as any weekly
counters. The result is that fewer counters are hit per new goal added
(nice to reduce points of constraint).

This is also preparation for adding predictions, knowing that the
counters are named events will be helpful in the abscence of type
information stored with the goal. We can say "this is only named events"
instead of trying to figure it out later. Also avoids having
records/another format that gets stored with a goal.