Remove TODO
Using the Name for features store doesn't help all that much. The
configmap lib is only used for the features store, and the particular
name doesn't change, at least for now. The features part of the app is
also less of a major product feature anymore, and may not be worth
putting more time into in the short term