router: events for user should exclude global_counter
global_counter is a weird case, but in the list of events a user has
completed, it shouldn't be included. When a list of just strings is
passed, including global_counter seems wrong. If all counters are
included and the types can be checked, then it might normally be fine.

Previously including global_counter would break predictions as it was
seen as any other event, but the counter couldn't be found as there was
no named counter for global_counter.