features_store: Use global store lib
Instead of having separate configmap/file. Using the configmap was a
nice workaround for a highly available Kube app but feature flags is
less of what we want to do and configmaps require kubernetes.

Using the same storelib for everything simplifies operations / choices
when deploying our app. Feature flags are relatively small config wise
(KB level sizing) so even frequent access to S3 or the like is more
API-call bound. This can all be improved in the future, but right now
it's slowing things down to have this complication