counters: Add delay between startup of each counter
A previous run of the application failed running out of file descriptors
as too many processes were attempting to start at once.

This isn't a great solution to that particular problem, overall, but
allows for user-setting of the value that could help with this or other
startup problems.

Handling just the file descriptor problem would leave open the potential
for other problems that rate limiting can solve